Reintroduced the `Children` path node, and fixes for tests (#12)
I noticed that I was getting some strange results when playing around in
the REPL, and I found a couple of issues. Also, some of the tests were
passing unexpectedly, so I updated a few of the path-access definitions
in `XmlNode` and `XmlSeq` and corrected the tests.

I also added applyDynamic where it was missing. I'm still not 100% sure
what we have is correct, but all the tests are passing again, and my
limited testing in the REPL seems to be more consistent.

One invariant we should have now is that `xml.foo` and `xml.*` should
produce the same results, if xml only contains nodes called `foo`.