Shuffled implicits around so that implicit search inside macro works
I think the problem before was because the normal primitive implicit
serializers were in a companion object for a type which was not being
searched. I would need to spend longer with it to fully understand why
the implicits were visible when requesting an `XmlSeqSerializer[String]`
from the REPL, whilst not being visible when attempting to access them
from the macro.

It's possible (though I'm not completely sure of it) that the priorities
on the primitive implicits were different inside the macro, and each of
these implicits was visible to the compiler through two different
routes: once through one companion object, and once through the owner of
that companion object, and each one was clashing with itself and was
therefore ambiguous. This is entirely hypothetical, and there may be a
more obvious explanation, but I don't have it.

Anyway, the macro is closer to working now, so you should be able to
make some progress from here!